[mod] Adjust the path for all module [add] Add config_setting.py. Every file can import Parameters from it to access ByteTrack setting
