+++ lib/ByteTrack_yolov7/__init__.py
@@ -1,1 +1,2 @@
-from yolov7 import CLS2ID, ID2CLS, DEVICE
+from .yolov7 import CLS2ID, ID2CLS, DEVICE
+from .Yolov7_Track import  Yolov7_Track
