+++ lib/ByteTrack_yolov7/tracker/matching.py
@@ -5,8 +5,10 @@ import lap
 from scipy.spatial.distance import cdist
 
 from cython_bbox import bbox_overlaps as bbox_ious
-from . import kalman_filter
-import time
+import sys
+from pathlib import Path
+sys.path.append(str(Path(__file__).resolve().parent))
+import kalman_filter
 
 def merge_matches(m1, m2, shape):
     O,P,Q = shape
